Fix legacy registry rollback on skill save
