+++ src/commands/skill_cmds.rs
@@ -36,6 +36,13 @@ use super::helpers::{
 };
 use super::{App, CommandFailure};
 
+mod observed;
+mod save;
+mod shared;
+mod snapshot;
+
+use self::shared::*;
+
 impl App {
     pub fn cmd_add(
         &self,
@@ -666,1328 +673,5 @@ impl App {
             }),
             meta,
         ))
-    }
-
-    pub fn cmd_save(
-        &self,
-        args: &SaveArgs,
-        request_id: &str,
-    ) -> std::result::Result<(serde_json::Value, Meta), CommandFailure> {
-        validate_skill_name(&args.skill).map_err(map_arg)?;
-        let _workspace = self.ctx.lock_workspace().map_err(map_lock)?;
-        self.ensure_write_repo_ready()?;
-        let _lock = self.ctx.lock_skill(&args.skill).map_err(map_lock)?;
-        let skill_rel = format!("skills/{}", args.skill);
-        let skill_path = self.ctx.root.join(&skill_rel);
-        if !skill_path.exists() {
-            return Err(CommandFailure::new(
-                ErrorCode::SkillNotFound,
-                format!("skill '{}' not found", args.skill),
-            ));
-        }
-
-        gitops::stage_path(&self.ctx, Path::new(&skill_rel)).map_err(map_git)?;
-        let changed = gitops::has_staged_changes_for_path(&self.ctx, Path::new(&skill_rel))
-            .map_err(map_git)?;
-        if !changed {
-            return Ok((json!({"skill": args.skill, "noop": true}), Meta::default()));
-        }
-
-        let paths = RegistryStatePaths::from_app_context(&self.ctx);
-        let had_registry_layout = paths.registry_dir.exists();
-        let had_legacy_layout = paths.legacy_state_dir_exists();
-        paths.ensure_layout().map_err(map_registry_state)?;
-        let registry_backup = snapshot_registry_audit_state(&paths).map_err(map_registry_state)?;
-        let message = args
-            .message
-            .clone()
-            .unwrap_or_else(|| format!("save({}): event", args.skill));
-        let op_id = match record_registry_operation(
-            &paths,
-            "skill.save",
-            json!({
-                "skill": args.skill,
-                "request_id": request_id
-            }),
-            json!({
-                "noop": false
-            }),
-        ) {
-            Ok(op_id) => op_id,
-            Err(err) => {
-                let _ =
-                    gitops::run_git_allow_failure(&self.ctx, &["reset", "HEAD", "--", &skill_rel]);
-                rollback_registry_audit_after_failure(
-                    &self.ctx,
-                    &paths,
-                    &registry_backup,
-                    had_registry_layout,
-                    had_legacy_layout,
-                );
-                return Err(map_registry_state(err));
-            }
-        };
-        if let Err(err) = maybe_skill_fault("skill_save_after_operation") {
-            let _ = gitops::run_git_allow_failure(&self.ctx, &["reset", "HEAD", "--", &skill_rel]);
-            rollback_registry_audit_after_failure(
-                &self.ctx,
-                &paths,
-                &registry_backup,
-                had_registry_layout,
-                had_legacy_layout,
-            );
-            return Err(err);
-        }
-        if let Err(err) = stage_registry_state(&self.ctx, &paths) {
-            let _ = gitops::run_git_allow_failure(&self.ctx, &["reset", "HEAD", "--", &skill_rel]);
-            rollback_registry_audit_after_failure(
-                &self.ctx,
-                &paths,
-                &registry_backup,
-                had_registry_layout,
-                had_legacy_layout,
-            );
-            return Err(err);
-        }
-        let commit = match gitops::commit(&self.ctx, &message) {
-            Ok(commit) => commit,
-            Err(err) => {
-                let _ =
-                    gitops::run_git_allow_failure(&self.ctx, &["reset", "HEAD", "--", &skill_rel]);
-                rollback_registry_audit_after_failure(
-                    &self.ctx,
-                    &paths,
-                    &registry_backup,
-                    had_registry_layout,
-                    had_legacy_layout,
-                );
-                return Err(map_git(err));
-            }
-        };
-        let mut meta = Meta {
-            op_id: Some(op_id),
-            ..Meta::default()
-        };
-        maybe_autosync_or_queue(
-            &self.ctx,
-            "save",
-            request_id,
-            json!({"skill": args.skill, "commit": commit}),
-            &mut meta,
-        )?;
-
-        Ok((
-            json!({"skill": args.skill, "commit": commit, "noop": false}),
-            meta,
-        ))
-    }
-
-    pub fn cmd_import_observed(
-        &self,
-        args: &ImportObservedArgs,
-        request_id: &str,
-    ) -> std::result::Result<(serde_json::Value, Meta), CommandFailure> {
-        let _workspace = self.ctx.lock_workspace().map_err(map_lock)?;
-        self.ensure_write_repo_ready()?;
-        let paths = self.ensure_registry_layout()?;
-        let snapshot = paths.load_snapshot().map_err(map_registry_state)?;
-
-        let targets = observed_import_targets(&snapshot.targets.targets, args.target.as_deref())?;
-        let staging_root = self
-            .ctx
-            .state_dir
-            .join(format!("tmp-import-observed-{}", Uuid::new_v4()));
-        let cleanup_staging = || {
-            let _ = remove_path_if_exists(&staging_root);
-        };
-
-        remove_path_if_exists(&staging_root).map_err(map_io)?;
-        fs::create_dir_all(&staging_root).map_err(map_io)?;
-
-        let mut imported = Vec::new();
-        let mut skipped = Vec::new();
-        let mut imported_rels = Vec::new();
-
-        for target in targets {
-            let target_path = PathBuf::from(&target.path);
-            if !target_path.exists() {
-                skipped.push(json!({
-                    "target_id": target.target_id,
-                    "path": target.path,
-                    "reason": "target-missing",
-                }));
-                continue;
-            }
-            if !target_path.is_dir() {
-                skipped.push(json!({
-                    "target_id": target.target_id,
-                    "path": target.path,
-                    "reason": "target-not-directory",
-                }));
-                continue;
-            }
-
-            let mut entries = match fs::read_dir(&target_path) {
-                Ok(entries) => entries
-                    .filter_map(|entry| match entry {
-                        Ok(entry) => Some(entry),
-                        Err(err) => {
-                            skipped.push(json!({
-                                "target_id": target.target_id,
-                                "path": target.path,
-                                "reason": "entry-read-failed",
-                                "error": err.to_string(),
-                            }));
-                            None
-                        }
-                    })
-                    .collect::<Vec<_>>(),
-                Err(err) => {
-                    skipped.push(json!({
-                        "target_id": target.target_id,
-                        "path": target.path,
-                        "reason": "target-read-failed",
-                        "error": err.to_string(),
-                    }));
-                    continue;
-                }
-            };
-            entries.sort_by_key(|entry| entry.file_name());
-
-            for entry in entries {
-                let source_path = entry.path();
-                let file_type = match entry.file_type() {
-                    Ok(file_type) => file_type,
-                    Err(err) => {
-                        skipped.push(json!({
-                            "target_id": target.target_id,
-                            "source": source_path.display().to_string(),
-                            "reason": "file-type-failed",
-                            "error": err.to_string(),
-                        }));
-                        continue;
-                    }
-                };
-                let (copy_source, source_kind, resolved_source) = match observed_skill_copy_source(
-                    &source_path,
-                    &file_type,
-                    &mut skipped,
-                    &target,
-                ) {
-                    Some(source) => source,
-                    None => continue,
-                };
-                if !has_skill_entrypoint(&copy_source) {
-                    continue;
-                }
-
-                let skill_id = match entry.file_name().into_string() {
-                    Ok(name) => name,
-                    Err(name) => {
-                        skipped.push(json!({
-                            "target_id": target.target_id,
-                            "source": source_path.display().to_string(),
-                            "name": name.to_string_lossy(),
-                            "reason": "non-utf8-name",
-                        }));
-                        continue;
-                    }
-                };
-
-                if let Err(err) = validate_skill_name(&skill_id) {
-                    skipped.push(json!({
-                        "target_id": target.target_id,
-                        "skill": skill_id,
-                        "source": source_path.display().to_string(),
-                        "reason": "invalid-skill-name",
-                        "error": err.to_string(),
-                    }));
-                    continue;
-                }
-
-                let dst = self.ctx.skill_path(&skill_id);
-                if dst.exists() {
-                    skipped.push(json!({
-                        "target_id": target.target_id,
-                        "skill": skill_id,
-                        "source": source_path.display().to_string(),
-                        "reason": "already-exists",
-                    }));
-                    continue;
-                }
-
-                let staging_skill = staging_root.join(&skill_id);
-                let _ = remove_path_if_exists(&staging_skill);
-                match copy_dir_recursive_without_symlinks(&copy_source, &staging_skill) {
-                    Ok(()) => {}
-                    Err(err) => {
-                        let _ = remove_path_if_exists(&staging_skill);
-                        skipped.push(json!({
-                            "target_id": target.target_id,
-                            "skill": skill_id,
-                            "source": source_path.display().to_string(),
-                            "reason": "copy-failed",
-                            "error": err.to_string(),
-                        }));
-                        continue;
-                    }
-                }
-
-                if let Err(err) = fs::rename(&staging_skill, &dst) {
-                    cleanup_staging();
-                    rollback_imported_skills(&self.ctx, &imported_rels);
-                    return Err(map_io(err));
-                }
-
-                let skill_rel = format!("skills/{}", skill_id);
-                if let Err(err) = gitops::stage_path(&self.ctx, Path::new(&skill_rel)) {
-                    cleanup_staging();
-                    rollback_imported_skills(&self.ctx, &imported_rels);
-                    rollback_added_skill(&self.ctx, &skill_rel, &dst);
-                    return Err(map_git(err));
-                }
-                imported_rels.push(skill_rel);
-                let mut imported_item = json!({
-                    "target_id": target.target_id,
-                    "skill": skill_id,
-                    "source": source_path.display().to_string(),
-                    "source_kind": source_kind,
-                    "path": dst.display().to_string(),
-                });
-                if let Some(resolved_source) = resolved_source {
-                    imported_item["resolved_source"] = json!(resolved_source);
-                }
-                imported.push(imported_item);
-            }
-        }
-
-        cleanup_staging();
-
-        let mut meta = Meta::default();
-        let previous_head = gitops::head(&self.ctx).map_err(map_git)?;
-        let registry_backup = snapshot_registry_audit_state(&paths).map_err(map_registry_state)?;
-        let mut changed = false;
-        for skill_rel in &imported_rels {
-            match gitops::has_staged_changes_for_path(&self.ctx, Path::new(skill_rel)) {
-                Ok(true) => {
-                    changed = true;
-                    break;
-                }
-                Ok(false) => {}
-                Err(err) => {
-                    rollback_imported_skills(&self.ctx, &imported_rels);
-                    return Err(map_git(err));
-                }
-            }
-        }
-
-        let commit = if changed {
-            let message = match imported.len() {
-                1 => format!(
-                    "import-observed({}): from observed target",
-                    imported[0]["skill"].as_str().unwrap_or("skill")
-                ),
-                count => format!("import-observed: {} skills", count),
-            };
-            let commit = match gitops::commit(&self.ctx, &message) {
-                Ok(commit) => commit,
-                Err(err) => {
-                    rollback_imported_skills(&self.ctx, &imported_rels);
-                    return Err(map_git(err));
-                }
-            };
-            let post_commit = (|| -> std::result::Result<Meta, CommandFailure> {
-                let op_id = record_registry_operation(
-                    &paths,
-                    "skill.import_observed",
-                    json!({
-                        "target": args.target,
-                        "request_id": request_id
-                    }),
-                    json!({
-                        "commit": commit,
-                        "imported": imported,
-                        "skipped": skipped
-                    }),
-                )
-                .map_err(map_registry_state)?;
-                let state_commit =
-                    commit_registry_state(&self.ctx, "import-observed: record registry state")?;
-                let mut meta = Meta {
-                    op_id: Some(op_id),
-                    ..Meta::default()
-                };
-                maybe_autosync_or_queue(
-                    &self.ctx,
-                    "import-observed",
-                    request_id,
-                    json!({"commit": commit, "state_commit": state_commit, "count": imported.len()}),
-                    &mut meta,
-                )?;
-                Ok(meta)
-            })();
-            let post_meta = match post_commit {
-                Ok(result) => result,
-                Err(err) => {
-                    rollback_import_after_commit(
-                        &self.ctx,
-                        &paths,
-                        &registry_backup,
-                        &previous_head,
-                        &imported_rels,
-                    );
-                    return Err(err);
-                }
-            };
-            meta = post_meta;
-            Some(commit)
-        } else {
-            None
-        };
-
-        Ok((
-            json!({
-                "count": imported.len(),
-                "imported": imported,
-                "skipped": skipped,
-                "commit": commit,
-                "noop": !changed,
-            }),
-            meta,
-        ))
-    }
-
-    pub fn cmd_monitor_observed(
-        &self,
-        args: &MonitorObservedArgs,
-        request_id: &str,
-    ) -> std::result::Result<(serde_json::Value, Meta), CommandFailure> {
-        if !args.once && args.interval_seconds == 0 {
-            return Err(CommandFailure::new(
-                ErrorCode::ArgInvalid,
-                "--interval-seconds must be greater than 0 for long-running monitoring",
-            ));
-        }
-
-        let mut cycles = 0_u64;
-        let mut totals = MonitorTotals::default();
-        let mut last_cycle = json!(null);
-        let mut meta = Meta::default();
-
-        loop {
-            let (cycle, cycle_meta) = self.monitor_observed_once(args, request_id)?;
-            cycles += 1;
-            totals.add_cycle(&cycle);
-            last_cycle = cycle;
-            merge_monitor_meta(&mut meta, cycle_meta);
-
-            if args.once || args.max_cycles.is_some_and(|max| cycles >= max) {
-                break;
-            }
-
-            thread::sleep(Duration::from_secs(args.interval_seconds));
-        }
-
-        Ok((
-            json!({
-                "cycles": cycles,
-                "totals": totals.to_json(),
-                "last_cycle": last_cycle,
-            }),
-            meta,
-        ))
-    }
-
-    fn monitor_observed_once(
-        &self,
-        args: &MonitorObservedArgs,
-        request_id: &str,
-    ) -> std::result::Result<(serde_json::Value, Meta), CommandFailure> {
-        let _workspace = self.ctx.lock_workspace().map_err(map_lock)?;
-        self.ensure_write_repo_ready()?;
-        let paths = self.ensure_registry_layout()?;
-        let snapshot = paths.load_snapshot().map_err(map_registry_state)?;
-        let targets = observed_import_targets(&snapshot.targets.targets, args.target.as_deref())?;
-        let staging_root = self
-            .ctx
-            .state_dir
-            .join(format!("tmp-monitor-observed-{}", Uuid::new_v4()));
-        let cleanup_staging = || {
-            let _ = remove_path_if_exists(&staging_root);
-        };
-
-        remove_path_if_exists(&staging_root).map_err(map_io)?;
-        fs::create_dir_all(&staging_root).map_err(map_io)?;
-
-        let mut imported = Vec::new();
-        let mut updated = Vec::new();
-        let mut skipped = Vec::new();
-        let mut unchanged_count = 0_usize;
-        let mut changed_rels = Vec::new();
-        let mut imported_rels = Vec::new();
-        let mut update_rollbacks = Vec::new();
-        let mut seen_skill_ids = BTreeSet::new();
-
-        for target in targets {
-            let target_path = PathBuf::from(&target.path);
-            if !target_path.exists() {
-                skipped.push(json!({
-                    "target_id": target.target_id,
-                    "path": target.path,
-                    "reason": "target-missing",
-                }));
-                continue;
-            }
-            if !target_path.is_dir() {
-                skipped.push(json!({
-                    "target_id": target.target_id,
-                    "path": target.path,
-                    "reason": "target-not-directory",
-                }));
-                continue;
-            }
-
-            let mut entries = match fs::read_dir(&target_path) {
-                Ok(entries) => entries
-                    .filter_map(|entry| match entry {
-                        Ok(entry) => Some(entry),
-                        Err(err) => {
-                            skipped.push(json!({
-                                "target_id": target.target_id,
-                                "path": target.path,
-                                "reason": "entry-read-failed",
-                                "error": err.to_string(),
-                            }));
-                            None
-                        }
-                    })
-                    .collect::<Vec<_>>(),
-                Err(err) => {
-                    skipped.push(json!({
-                        "target_id": target.target_id,
-                        "path": target.path,
-                        "reason": "target-read-failed",
-                        "error": err.to_string(),
-                    }));
-                    continue;
-                }
-            };
-            entries.sort_by_key(|entry| entry.file_name());
-
-            for entry in entries {
-                let source_path = entry.path();
-                let file_type = match entry.file_type() {
-                    Ok(file_type) => file_type,
-                    Err(err) => {
-                        skipped.push(json!({
-                            "target_id": target.target_id,
-                            "source": source_path.display().to_string(),
-                            "reason": "file-type-failed",
-                            "error": err.to_string(),
-                        }));
-                        continue;
-                    }
-                };
-                let (copy_source, source_kind, resolved_source) = match observed_skill_copy_source(
-                    &source_path,
-                    &file_type,
-                    &mut skipped,
-                    &target,
-                ) {
-                    Some(source) => source,
-                    None => continue,
-                };
-                if !has_skill_entrypoint(&copy_source) {
-                    continue;
-                }
-
-                let skill_id = match entry.file_name().into_string() {
-                    Ok(name) => name,
-                    Err(name) => {
-                        skipped.push(json!({
-                            "target_id": target.target_id,
-                            "source": source_path.display().to_string(),
-                            "name": name.to_string_lossy(),
-                            "reason": "non-utf8-name",
-                        }));
-                        continue;
-                    }
-                };
-
-                if let Err(err) = validate_skill_name(&skill_id) {
-                    skipped.push(json!({
-                        "target_id": target.target_id,
-                        "skill": skill_id,
-                        "source": source_path.display().to_string(),
-                        "reason": "invalid-skill-name",
-                        "error": err.to_string(),
-                    }));
-                    continue;
-                }
-
-                if !seen_skill_ids.insert(skill_id.clone()) {
-                    skipped.push(json!({
-                        "target_id": target.target_id,
-                        "skill": skill_id,
-                        "source": source_path.display().to_string(),
-                        "reason": "duplicate-observed-skill",
-                    }));
-                    continue;
-                }
-
-                let staging_skill = staging_root.join("next").join(&skill_id);
-                let _ = remove_path_if_exists(&staging_skill);
-                match copy_dir_recursive_without_symlinks(&copy_source, &staging_skill) {
-                    Ok(()) => {}
-                    Err(err) => {
-                        let _ = remove_path_if_exists(&staging_skill);
-                        skipped.push(json!({
-                            "target_id": target.target_id,
-                            "skill": skill_id,
-                            "source": source_path.display().to_string(),
-                            "reason": "copy-failed",
-                            "error": err.to_string(),
-                        }));
-                        continue;
-                    }
-                }
-
-                let dst = self.ctx.skill_path(&skill_id);
-                let skill_rel = format!("skills/{}", skill_id);
-                let mut item = json!({
-                    "target_id": target.target_id,
-                    "skill": skill_id,
-                    "source": source_path.display().to_string(),
-                    "source_kind": source_kind,
-                    "path": dst.display().to_string(),
-                });
-                if let Some(resolved_source) = resolved_source {
-                    item["resolved_source"] = json!(resolved_source);
-                }
-
-                if dst.exists() {
-                    match materialized_dirs_equal(&dst, &staging_skill) {
-                        Ok(true) => {
-                            unchanged_count += 1;
-                            let _ = remove_path_if_exists(&staging_skill);
-                            continue;
-                        }
-                        Ok(false) => {}
-                        Err(err) => {
-                            let _ = remove_path_if_exists(&staging_skill);
-                            skipped.push(json!({
-                                "target_id": target.target_id,
-                                "skill": item["skill"].clone(),
-                                "source": source_path.display().to_string(),
-                                "reason": "compare-failed",
-                                "error": err.to_string(),
-                            }));
-                            continue;
-                        }
-                    }
-
-                    let previous = staging_root.join("previous").join(
-                        item["skill"]
-                            .as_str()
-                            .expect("monitor item skill is string"),
-                    );
-                    if let Some(parent) = previous.parent() {
-                        fs::create_dir_all(parent).map_err(map_io)?;
-                    }
-                    if let Err(err) = fs::rename(&dst, &previous) {
-                        rollback_monitor_changes(&self.ctx, &imported_rels, &update_rollbacks);
-                        cleanup_staging();
-                        return Err(map_io(err));
-                    }
-                    if let Err(err) = fs::rename(&staging_skill, &dst) {
-                        let _ = fs::rename(&previous, &dst);
-                        rollback_monitor_changes(&self.ctx, &imported_rels, &update_rollbacks);
-                        cleanup_staging();
-                        return Err(map_io(err));
-                    }
-                    if let Err(err) = gitops::stage_path(&self.ctx, Path::new(&skill_rel)) {
-                        let _ = remove_path_if_exists(&dst);
-                        let _ = fs::rename(&previous, &dst);
-                        rollback_monitor_changes(&self.ctx, &imported_rels, &update_rollbacks);
-                        cleanup_staging();
-                        return Err(map_git(err));
-                    }
-                    update_rollbacks.push(MonitorUpdateRollback {
-                        skill_rel: skill_rel.clone(),
-                        dst,
-                        previous,
-                    });
-                    changed_rels.push(skill_rel);
-                    updated.push(item);
-                } else {
-                    if let Err(err) = fs::rename(&staging_skill, &dst) {
-                        rollback_monitor_changes(&self.ctx, &imported_rels, &update_rollbacks);
-                        cleanup_staging();
-                        return Err(map_io(err));
-                    }
-                    if let Err(err) = gitops::stage_path(&self.ctx, Path::new(&skill_rel)) {
-                        rollback_added_skill(&self.ctx, &skill_rel, &dst);
-                        rollback_monitor_changes(&self.ctx, &imported_rels, &update_rollbacks);
-                        cleanup_staging();
-                        return Err(map_git(err));
-                    }
-                    imported_rels.push(skill_rel.clone());
-                    changed_rels.push(skill_rel);
-                    imported.push(item);
-                }
-            }
-        }
-
-        let mut has_changes = false;
-        for skill_rel in &changed_rels {
-            match gitops::has_staged_changes_for_path(&self.ctx, Path::new(skill_rel)) {
-                Ok(true) => {
-                    has_changes = true;
-                    break;
-                }
-                Ok(false) => {}
-                Err(err) => {
-                    rollback_monitor_changes(&self.ctx, &imported_rels, &update_rollbacks);
-                    cleanup_staging();
-                    return Err(map_git(err));
-                }
-            }
-        }
-
-        let mut meta = Meta::default();
-        let previous_head = gitops::head(&self.ctx).map_err(map_git)?;
-        let registry_backup = snapshot_registry_audit_state(&paths).map_err(map_registry_state)?;
-        let commit = if has_changes {
-            let change_count = imported.len() + updated.len();
-            let message = if change_count == 1 {
-                let skill = imported
-                    .first()
-                    .or_else(|| updated.first())
-                    .and_then(|item| item["skill"].as_str())
-                    .unwrap_or("skill");
-                format!("monitor-observed({}): sync observed skill", skill)
-            } else {
-                format!("monitor-observed: {} skills", change_count)
-            };
-            let commit = match gitops::commit(&self.ctx, &message) {
-                Ok(commit) => commit,
-                Err(err) => {
-                    rollback_monitor_changes(&self.ctx, &imported_rels, &update_rollbacks);
-                    cleanup_staging();
-                    return Err(map_git(err));
-                }
-            };
-            let post_commit = (|| -> std::result::Result<Meta, CommandFailure> {
-                let op_id = record_registry_operation(
-                    &paths,
-                    "skill.monitor_observed",
-                    json!({
-                        "target": args.target,
-                        "request_id": request_id
-                    }),
-                    json!({
-                        "commit": commit,
-                        "imported": imported,
-                        "updated": updated,
-                        "skipped": skipped,
-                        "unchanged_count": unchanged_count,
-                    }),
-                )
-                .map_err(map_registry_state)?;
-                record_observed_skill_events(
-                    &paths,
-                    &snapshot.projections.projections,
-                    imported.iter().chain(updated.iter()),
-                    &commit,
-                )
-                .map_err(map_registry_state)?;
-                maybe_skill_fault("skill_monitor_after_observation")?;
-                let state_commit =
-                    commit_registry_state(&self.ctx, "monitor-observed: record registry state")?;
-                let mut meta = Meta {
-                    op_id: Some(op_id),
-                    ..Meta::default()
-                };
-                maybe_autosync_or_queue(
-                    &self.ctx,
-                    "monitor-observed",
-                    request_id,
-                    json!({
-                        "commit": commit,
-                        "state_commit": state_commit,
-                        "imported": imported.len(),
-                        "updated": updated.len(),
-                    }),
-                    &mut meta,
-                )?;
-                Ok(meta)
-            })();
-            let post_meta = match post_commit {
-                Ok(result) => result,
-                Err(err) => {
-                    rollback_monitor_after_commit(
-                        &self.ctx,
-                        &paths,
-                        &registry_backup,
-                        &previous_head,
-                        &imported_rels,
-                        &update_rollbacks,
-                    );
-                    cleanup_staging();
-                    return Err(err);
-                }
-            };
-            meta = post_meta;
-            Some(commit)
-        } else {
-            None
-        };
-
-        cleanup_staging();
-        let change_count = imported.len() + updated.len();
-        Ok((
-            json!({
-                "count": change_count,
-                "imported_count": imported.len(),
-                "updated_count": updated.len(),
-                "unchanged_count": unchanged_count,
-                "skipped_count": skipped.len(),
-                "imported": imported,
-                "updated": updated,
-                "skipped": skipped,
-                "commit": commit,
-                "noop": !has_changes,
-            }),
-            meta,
-        ))
-    }
-
-    pub fn cmd_snapshot(
-        &self,
-        args: &SkillOnlyArgs,
-        request_id: &str,
-    ) -> std::result::Result<(serde_json::Value, Meta), CommandFailure> {
-        validate_skill_name(&args.skill).map_err(map_arg)?;
-        let _workspace = self.ctx.lock_workspace().map_err(map_lock)?;
-        self.ensure_write_repo_ready()?;
-        ensure_skill_exists(&self.ctx, &args.skill)?;
-        let _lock = self.ctx.lock_skill(&args.skill).map_err(map_lock)?;
-
-        let short = gitops::short_head(&self.ctx).map_err(map_git)?;
-        let ts = Utc::now().format("%Y%m%dT%H%M%S%fZ");
-        let tag = format!("snapshot/{}/{}-{}", args.skill, ts, short);
-        gitops::create_annotated_tag(&self.ctx, &tag, &format!("snapshot {}", args.skill))
-            .map_err(map_git)?;
-        if let Err(err) = gitops::append_history_audit_event(
-            &self.ctx,
-            "skill.snapshot",
-            json!({"skill": args.skill, "tag": tag.clone()}),
-            request_id,
-        ) {
-            let mut failure = map_git(err);
-            let rollback_output = gitops::run_git_allow_failure(&self.ctx, &["tag", "-d", &tag]);
-            let mut rollback_errors = Vec::new();
-            match rollback_output {
-                Ok(output) if output.status.success() => {}
-                Ok(output) => rollback_errors.push(json!({
-                    "step": "delete_snapshot_tag",
-                    "message": String::from_utf8_lossy(&output.stderr).trim().to_string()
-                })),
-                Err(err) => rollback_errors.push(json!({
-                    "step": "delete_snapshot_tag",
-                    "message": err.to_string()
-                })),
-            }
-            if !rollback_errors.is_empty() {
-                failure.details = json!({ "rollback_errors": rollback_errors });
-            }
-            return Err(failure);
-        }
-
-        let mut meta = Meta::default();
-        maybe_autosync_or_queue(
-            &self.ctx,
-            "snapshot",
-            request_id,
-            json!({"skill": args.skill, "tag": tag}),
-            &mut meta,
-        )?;
-
-        Ok((json!({"skill": args.skill, "tag": tag}), meta))
-    }
-}
-
-fn maybe_skill_fault(tag: &str) -> std::result::Result<(), CommandFailure> {
-    if std::env::var("LOOM_FAULT_INJECT").ok().as_deref() == Some(tag) {
-        return Err(CommandFailure::new(
-            ErrorCode::InternalError,
-            format!("fault injected at {}", tag),
-        ));
-    }
-    Ok(())
-}
-
-fn rollback_fault_active(tag: &str) -> bool {
-    std::env::var("LOOM_ROLLBACK_FAULT_INJECT").ok().as_deref() == Some(tag)
-}
-
-fn push_rollback_error(errors: &mut Vec<Value>, step: &str, message: impl ToString) {
-    errors.push(json!({
-        "step": step,
-        "message": message.to_string(),
-    }));
-}
-
-fn maybe_push_rollback_fault(errors: &mut Vec<Value>, step: &str) -> bool {
-    if rollback_fault_active(step) {
-        push_rollback_error(errors, step, format!("fault injected at {}", step));
-        return true;
-    }
-    false
-}
-
-fn rollback_project_mutation(
-    paths: &RegistryStatePaths,
-    materialized_path: &Path,
-    backup: Option<&serde_json::Value>,
-    original_bindings: &RegistryBindingsFile,
-    original_rules: &RegistryRulesFile,
-    original_projections: &RegistryProjectionsFile,
-) -> Vec<Value> {
-    let mut errors = Vec::new();
-    if let Some(backup) = backup {
-        if !maybe_push_rollback_fault(&mut errors, "restore_projection_path")
-            && let Err(err) = restore_path_from_backup(materialized_path, backup)
-        {
-            push_rollback_error(&mut errors, "restore_projection_path", err);
-        }
-    } else {
-        if !maybe_push_rollback_fault(&mut errors, "remove_projection_path")
-            && let Err(err) = remove_path_if_exists(materialized_path)
-        {
-            push_rollback_error(&mut errors, "remove_projection_path", err);
-        }
-    }
-    if let Err(err) = rollback_registry_state(
-        paths,
-        original_bindings,
-        original_rules,
-        original_projections,
-    ) {
-        push_rollback_error(&mut errors, "restore_registry_state", err);
-    }
-    errors
-}
-
-fn rollback_capture_mutation(
-    ctx: &crate::state::AppContext,
-    skill_path: &Path,
-    source_backup: Option<&serde_json::Value>,
-    source_replaced: bool,
-    previous_head: &str,
-    previous_index: &gitops::IndexSnapshot,
-    commit_created: bool,
-) -> Vec<Value> {
-    let mut errors = Vec::new();
-    if commit_created {
-        // Preserve worktree content while removing only the command-created commit.
-        if !maybe_push_rollback_fault(&mut errors, "reset_command_created_commit") {
-            match gitops::run_git_allow_failure(ctx, &["reset", "--soft", previous_head]) {
-                Ok(output) if output.status.success() => {}
-                Ok(output) => push_rollback_error(
-                    &mut errors,
-                    "reset_command_created_commit",
-                    String::from_utf8_lossy(&output.stderr).trim(),
-                ),
-                Err(err) => push_rollback_error(&mut errors, "reset_command_created_commit", err),
-            }
-        }
-    }
-
-    if source_replaced {
-        if let Some(backup) = source_backup {
-            if !maybe_push_rollback_fault(&mut errors, "restore_source_path")
-                && let Err(err) = restore_path_from_backup(skill_path, backup)
-            {
-                push_rollback_error(&mut errors, "restore_source_path", err);
-            }
-        } else {
-            if !maybe_push_rollback_fault(&mut errors, "remove_source_path")
-                && let Err(err) = remove_path_if_exists(skill_path)
-            {
-                push_rollback_error(&mut errors, "remove_source_path", err);
-            }
-        }
-    }
-
-    if !maybe_push_rollback_fault(&mut errors, "restore_git_index")
-        && let Err(err) = gitops::restore_index(ctx, previous_index)
-    {
-        push_rollback_error(&mut errors, "restore_git_index", err);
-    }
-    errors
-}
-
-fn ensure_capture_source_not_drifted(
-    ctx: &crate::state::AppContext,
-    projection: &RegistryProjectionInstance,
-    skill_rel: &Path,
-) -> std::result::Result<(), CommandFailure> {
-    let skill_rel_str = skill_rel.to_string_lossy();
-    let committed = git_diff_has_changes(
-        ctx,
-        &[&projection.last_applied_rev, "HEAD", "--", &skill_rel_str],
-    )?;
-    let staged = git_diff_has_changes(ctx, &["--cached", "--", &skill_rel_str])?;
-    let unstaged = git_diff_has_changes(ctx, &["--", &skill_rel_str])?;
-
-    if !(committed || staged || unstaged) {
-        return Ok(());
-    }
-
-    let current_rev = gitops::head(ctx).map_err(map_git)?;
-    let mut failure = CommandFailure::new(
-        ErrorCode::CaptureConflict,
-        format!(
-            "source skill '{}' changed since projection '{}'; save or rollback source changes before capture",
-            projection.skill_id, projection.instance_id
-        ),
-    );
-    failure.details = json!({
-        "skill_id": projection.skill_id,
-        "instance_id": projection.instance_id,
-        "source_path": skill_rel.display().to_string(),
-        "last_applied_rev": projection.last_applied_rev,
-        "current_rev": current_rev,
-        "committed": committed,
-        "staged": staged,
-        "unstaged": unstaged
-    });
-    Err(failure)
-}
-
-fn git_diff_has_changes(
-    ctx: &crate::state::AppContext,
-    args: &[&str],
-) -> std::result::Result<bool, CommandFailure> {
-    let mut full_args = vec!["diff", "--quiet"];
-    full_args.extend(args.iter().copied());
-    let output = gitops::run_git_allow_failure(ctx, &full_args).map_err(map_git)?;
-    match output.status.code() {
-        Some(0) => Ok(false),
-        Some(1) => Ok(true),
-        _ => Err(map_git(anyhow::anyhow!(
-            "git {:?} failed: {}",
-            full_args,
-            String::from_utf8_lossy(&output.stderr).trim()
-        ))),
-    }
-}
-
-fn rollback_registry_state(
-    paths: &RegistryStatePaths,
-    original_bindings: &RegistryBindingsFile,
-    original_rules: &RegistryRulesFile,
-    original_projections: &RegistryProjectionsFile,
-) -> anyhow::Result<()> {
-    if rollback_fault_active("restore_registry_state") {
-        anyhow::bail!("fault injected at restore_registry_state");
-    }
-    paths.save_bindings_rules_projections(
-        original_bindings,
-        original_rules,
-        original_projections,
-    )?;
-    Ok(())
-}
-
-#[derive(Debug, Default)]
-struct MonitorTotals {
-    imported: usize,
-    updated: usize,
-    unchanged: usize,
-    skipped: usize,
-}
-
-impl MonitorTotals {
-    fn add_cycle(&mut self, cycle: &serde_json::Value) {
-        self.imported += cycle["imported_count"].as_u64().unwrap_or(0) as usize;
-        self.updated += cycle["updated_count"].as_u64().unwrap_or(0) as usize;
-        self.unchanged += cycle["unchanged_count"].as_u64().unwrap_or(0) as usize;
-        self.skipped += cycle["skipped_count"].as_u64().unwrap_or(0) as usize;
-    }
-
-    fn to_json(&self) -> serde_json::Value {
-        json!({
-            "imported": self.imported,
-            "updated": self.updated,
-            "unchanged": self.unchanged,
-            "skipped": self.skipped,
-            "changed": self.imported + self.updated,
-        })
-    }
-}
-
-#[derive(Debug)]
-struct MonitorUpdateRollback {
-    skill_rel: String,
-    dst: PathBuf,
-    previous: PathBuf,
-}
-
-fn merge_monitor_meta(meta: &mut Meta, cycle_meta: Meta) {
-    if cycle_meta.op_id.is_some() {
-        meta.op_id = cycle_meta.op_id;
-    }
-    if cycle_meta.sync_state.is_some() {
-        meta.sync_state = cycle_meta.sync_state;
-    }
-    meta.warnings.extend(cycle_meta.warnings);
-}
-
-fn record_observed_skill_events<'a>(
-    paths: &RegistryStatePaths,
-    projections: &[RegistryProjectionInstance],
-    changes: impl Iterator<Item = &'a serde_json::Value>,
-    commit: &str,
-) -> anyhow::Result<()> {
-    for item in changes {
-        let Some(skill_id) = item["skill"].as_str() else {
-            continue;
-        };
-        let path = item["source"]
-            .as_str()
-            .or_else(|| item["path"].as_str())
-            .map(str::to_string);
-        for projection in projections.iter().filter(|p| p.skill_id == skill_id) {
-            record_registry_observation(
-                paths,
-                &projection.instance_id,
-                "monitor",
-                path.clone(),
-                None,
-                Some(commit.to_string()),
-            )?;
-        }
-    }
-    Ok(())
-}
-
-fn observed_import_targets(
-    targets: &[RegistryProjectionTarget],
-    target_id: Option<&str>,
-) -> std::result::Result<Vec<RegistryProjectionTarget>, CommandFailure> {
-    if let Some(target_id) = target_id {
-        let target = targets
-            .iter()
-            .find(|target| target.target_id == target_id)
-            .cloned()
-            .ok_or_else(|| {
-                CommandFailure::new(
-                    ErrorCode::TargetNotFound,
-                    format!("target '{}' not found", target_id),
-                )
-            })?;
-        if target.ownership != "observed" {
-            return Err(CommandFailure::new(
-                ErrorCode::ArgInvalid,
-                format!(
-                    "target '{}' has ownership '{}' and cannot be imported as observed",
-                    target.target_id, target.ownership
-                ),
-            ));
-        }
-        return Ok(vec![target]);
-    }
-
-    Ok(targets
-        .iter()
-        .filter(|target| target.ownership == "observed")
-        .cloned()
-        .collect())
-}
-
-fn observed_skill_copy_source(
-    source_path: &Path,
-    file_type: &fs::FileType,
-    skipped: &mut Vec<serde_json::Value>,
-    target: &RegistryProjectionTarget,
-) -> Option<(PathBuf, &'static str, Option<String>)> {
-    if file_type.is_dir() {
-        return Some((source_path.to_path_buf(), "directory", None));
-    }
-    if !file_type.is_symlink() {
-        return None;
-    }
-
-    let metadata = match fs::metadata(source_path) {
-        Ok(metadata) => metadata,
-        Err(err) => {
-            skipped.push(json!({
-                "target_id": target.target_id.clone(),
-                "source": source_path.display().to_string(),
-                "reason": "symlink-target-failed",
-                "error": err.to_string(),
-            }));
-            return None;
-        }
-    };
-    if !metadata.is_dir() {
-        return None;
-    }
-
-    match fs::canonicalize(source_path) {
-        Ok(resolved) => {
-            let display = resolved.display().to_string();
-            Some((resolved, "symlink", Some(display)))
-        }
-        Err(err) => {
-            skipped.push(json!({
-                "target_id": target.target_id.clone(),
-                "source": source_path.display().to_string(),
-                "reason": "symlink-resolve-failed",
-                "error": err.to_string(),
-            }));
-            None
-        }
-    }
-}
-
-fn has_skill_entrypoint(path: &Path) -> bool {
-    path.join("SKILL.md").is_file() || path.join("skill.md").is_file()
-}
-
-fn materialized_dirs_equal(left: &Path, right: &Path) -> anyhow::Result<bool> {
-    let left_files = collect_materialized_files(left)?;
-    let right_files = collect_materialized_files(right)?;
-    if left_files.len() != right_files.len() {
-        return Ok(false);
-    }
-
-    for (rel, left_body) in left_files {
-        match right_files.get(&rel) {
-            Some(right_body) if right_body == &left_body => {}
-            _ => return Ok(false),
-        }
-    }
-
-    Ok(true)
-}
-
-fn collect_materialized_files(root: &Path) -> anyhow::Result<BTreeMap<PathBuf, Vec<u8>>> {
-    let mut files = BTreeMap::new();
-
-    for entry in WalkDir::new(root).follow_links(false).into_iter() {
-        let entry = entry.with_context(|| format!("failed to walk {}", root.display()))?;
-        let rel = entry.path().strip_prefix(root).with_context(|| {
-            format!(
-                "failed to derive relative path for {} under {}",
-                entry.path().display(),
-                root.display()
-            )
-        })?;
-        if rel.as_os_str().is_empty() || entry.file_type().is_dir() {
-            continue;
-        }
-        if entry.file_type().is_symlink() {
-            return Ok(BTreeMap::from([(
-                rel.to_path_buf(),
-                b"__loom_symlink_marker__".to_vec(),
-            )]));
-        }
-        if entry.file_type().is_file() {
-            let body = fs::read(entry.path())
-                .with_context(|| format!("failed to read {}", entry.path().display()))?;
-            files.insert(rel.to_path_buf(), body);
-        }
-    }
-
-    Ok(files)
-}
-
-fn reset_command_created_commits(ctx: &crate::state::AppContext, previous_head: &str) {
-    let _ = gitops::run_git_allow_failure(ctx, &["reset", "--soft", previous_head]);
-}
-
-fn unstage_registry_state(ctx: &crate::state::AppContext) {
-    let _ = gitops::run_git_allow_failure(ctx, &["reset", "HEAD", "--", "state/registry"]);
-    let _ = gitops::run_git_allow_failure(ctx, &["reset", "HEAD", "--", "state/v3"]);
-}
-
-fn stage_registry_state(
-    ctx: &crate::state::AppContext,
-    paths: &RegistryStatePaths,
-) -> std::result::Result<(), CommandFailure> {
-    gitops::run_git(ctx, &["add", "-A", "--", "state/registry"]).map_err(map_git)?;
-    let legacy_v3_tracked =
-        gitops::run_git_allow_failure(ctx, &["ls-files", "--error-unmatch", "--", "state/v3"])
-            .map_err(map_git)?
-            .status
-            .success();
-    if paths.state_dir.join("v3").exists() || legacy_v3_tracked {
-        gitops::run_git(ctx, &["add", "-A", "--", "state/v3"]).map_err(map_git)?;
-    }
-    Ok(())
-}
-
-fn rollback_registry_audit_after_failure(
-    ctx: &crate::state::AppContext,
-    paths: &RegistryStatePaths,
-    registry_backup: &RegistryAuditStateBackup,
-    had_registry_layout: bool,
-    had_legacy_layout: bool,
-) {
-    let _ = restore_registry_audit_state(paths, registry_backup);
-    if !had_registry_layout && !had_legacy_layout {
-        let _ = remove_path_if_exists(&paths.registry_dir);
-    }
-    unstage_registry_state(ctx);
-}
-
-fn rollback_import_after_commit(
-    ctx: &crate::state::AppContext,
-    paths: &RegistryStatePaths,
-    registry_backup: &RegistryAuditStateBackup,
-    previous_head: &str,
-    imported_rels: &[String],
-) {
-    reset_command_created_commits(ctx, previous_head);
-    rollback_imported_skills(ctx, imported_rels);
-    let _ = restore_registry_audit_state(paths, registry_backup);
-    unstage_registry_state(ctx);
-}
-
-fn rollback_monitor_after_commit(
-    ctx: &crate::state::AppContext,
-    paths: &RegistryStatePaths,
-    registry_backup: &RegistryAuditStateBackup,
-    previous_head: &str,
-    imported_rels: &[String],
-    update_rollbacks: &[MonitorUpdateRollback],
-) {
-    reset_command_created_commits(ctx, previous_head);
-    rollback_monitor_changes(ctx, imported_rels, update_rollbacks);
-    let _ = restore_registry_audit_state(paths, registry_backup);
-    unstage_registry_state(ctx);
-}
-
-fn rollback_monitor_changes(
-    ctx: &crate::state::AppContext,
-    imported_rels: &[String],
-    update_rollbacks: &[MonitorUpdateRollback],
-) {
-    for update in update_rollbacks.iter().rev() {
-        let _ = remove_path_if_exists(&update.dst);
-        let _ = fs::rename(&update.previous, &update.dst);
-        let _ = gitops::run_git_allow_failure(ctx, &["reset", "HEAD", "--", &update.skill_rel]);
-    }
-
-    rollback_imported_skills(ctx, imported_rels);
-}
-
-fn rollback_imported_skills(ctx: &crate::state::AppContext, skill_rels: &[String]) {
-    for skill_rel in skill_rels {
-        let dst = ctx.root.join(skill_rel);
-        rollback_added_skill(ctx, skill_rel, &dst);
     }
 }
